gae_face_detection/demo/requirements.txt
# https://pypi.org/project/google-cloud-vision
google-cloud-vision==2.4.4

# https://pypi.org/project/Pillow
Pillow==8.3.2

# https://pypi.org/project/Flask
Flask==2.0.2
